feat: answer api 로직 수정
